+++ excel_toolkit_for_py/__init__.py
@@ -6,6 +6,8 @@ from .advanced_features import (add_chart, apply_conditional_formatting,
                                 extract_formulas, protect_excel,
                                 read_protected_excel, validate_empty_cells)
 from .conversions import csv_to_excel, excel_to_csv
+from .data_analysis import (calculate_basic_stats, calculate_correlations,
+                            create_pivot_table, detect_outliers)
 from .exporters import to_html, to_json, to_pdf, to_xml
 from .reader import read_csv, read_excel
 from .validations import validate_csv, validate_excel
@@ -33,5 +35,9 @@ __all__ = [
     'to_json',
     'to_xml',
     'to_html',
-    'to_pdf'
+    'to_pdf',
+    'calculate_basic_stats',
+    'detect_outliers',
+    'calculate_correlations',
+    'create_pivot_table'
 ]